Initial no-static
