+++ src/UnitTests/UnitTests.Net4.csproj
@@ -175,6 +175,7 @@
     <Compile Include="ConversionOperators.cs" />
     <Compile Include="CustomMapping.cs" />
     <Compile Include="EnumToNullableEnum.cs" />
+    <Compile Include="ExplicitMapperCreation.cs" />
     <Compile Include="ExpressionBridge.cs" />
     <Compile Include="ExpressionConversion.cs" />
     <Compile Include="ExtensionMethods.cs" />
